add flag: TW_IGNORE_MISC_WIPE_DATA
Some devices do not have a recovery key combo; they only have a wipe
data combo.  When the wipe data combo is used, the bootloader writes
a wipe command to /misc.  Newer versions of ASOP want to use /misc
for OTAs, so we need to mount /misc, but not wipe data when the
bootloader tries to.

This option creates a flag for those devices.  When the flag is
enabled, TWRP will ignore a --wipe_data command located in /misc
but will still run other commands written to /misc.

Change-Id: I78ceea05ec77f5aa4c7d0286cd7a50e5964290c6